move 'HeifContext::Image' class into its own file
