+++ libheif/api/libheif/api_structs.h
@@ -25,10 +25,11 @@
 #include "context.h"
 
 #include <memory>
+#include <codecs/image_item.h>
 
 struct heif_image_handle
 {
-  std::shared_ptr<HeifContext::Image> image;
+  std::shared_ptr<Image> image;
 
   // store reference to keep the context alive while we are using the handle (issue #147)
   std::shared_ptr<HeifContext> context;
