+++ libheif/codecs/mask_image.h
@@ -78,7 +78,7 @@ public:
                                  const std::shared_ptr<HeifPixelImage>& src_image,
                                  void* encoder_struct,
                                  const struct heif_encoding_options& options,
-                                 std::shared_ptr<HeifContext::Image>& out_image);
+                                 std::shared_ptr<Image>& out_image);
 };
 
 #endif //LIBHEIF_MASK_IMAGE_H
