+++ libheif/context.h
@@ -46,160 +46,8 @@ class HeifPixelImage;
 
 class StreamWriter;
 
-
-class ImageMetadata
-{
-public:
-  heif_item_id item_id;
-  std::string item_type;  // e.g. "Exif"
-  std::string content_type;
-  std::string item_uri_type;
-  std::vector<uint8_t> m_data;
-};
-
-
-class ImageGrid
-{
-public:
-  Error parse(const std::vector<uint8_t>& data);
-
-  std::vector<uint8_t> write() const;
-
-  std::string dump() const;
-
-  uint32_t get_width() const { return m_output_width; }
-
-  uint32_t get_height() const { return m_output_height; }
-
-  uint16_t get_rows() const
-  {
-    return m_rows;
-  }
-
-  uint16_t get_columns() const
-  {
-    return m_columns;
-  }
-
-  void set_num_tiles(uint16_t columns, uint16_t rows)
-  {
-    m_rows = rows;
-    m_columns = columns;
-  }
-
-  void set_output_size(uint32_t width, uint32_t height)
-  {
-    m_output_width = width;
-    m_output_height = height;
-  }
-
-private:
-  uint16_t m_rows = 0;
-  uint16_t m_columns = 0;
-  uint32_t m_output_width = 0;
-  uint32_t m_output_height = 0;
-};
-
-
-
-class ImageOverlay
-{
-public:
-  Error parse(size_t num_images, const std::vector<uint8_t>& data);
-
-  std::vector<uint8_t> write() const;
-
-  std::string dump() const;
-
-  void get_background_color(uint16_t col[4]) const;
-
-  uint32_t get_canvas_width() const { return m_width; }
-
-  uint32_t get_canvas_height() const { return m_height; }
-
-  size_t get_num_offsets() const { return m_offsets.size(); }
-
-  void get_offset(size_t image_index, int32_t* x, int32_t* y) const;
-
-  void set_background_color(const uint16_t rgba_color[4]) {
-    for (int i=0;i<4;i++) {
-      m_background_color[i] = rgba_color[i];
-    }
-  }
-
-  void set_canvas_size(uint32_t width, uint32_t height) { m_width = width; m_height = height; }
-
-  void add_image_on_top(heif_item_id image_id, int32_t offset_x, int32_t offset_y) {
-    m_offsets.emplace_back(ImageWithOffset{image_id, offset_x, offset_y});
-  }
-
-  struct ImageWithOffset
-  {
-    heif_item_id image_id;
-    int32_t x, y;
-  };
-
-  const std::vector<ImageWithOffset>& get_overlay_stack() const { return m_offsets; }
-
-private:
-  uint8_t m_version = 0;
-  uint8_t m_flags = 0;
-  uint16_t m_background_color[4] { 0,0,0,0 };
-  uint32_t m_width = 0;
-  uint32_t m_height = 0;
-
-  std::vector<ImageWithOffset> m_offsets;
-};
-
-
-#define TILD_OFFSET_NOT_AVAILABLE 0
-#define TILD_OFFSET_SEE_LOWER_RESOLUTION_LAYER 1
-#define TILD_OFFSET_NOT_LOADED 10
-
-class TildHeader
-{
-public:
-  void set_parameters(const heif_tild_image_parameters& params);
-
-  Error parse(size_t num_images, const std::vector<uint8_t>& data);
-
-  std::vector<uint8_t> write();
-
-  std::string dump() const;
-
-  uint64_t number_of_tiles() const;
-
-  uint64_t nTiles_h() const;
-
-  void set_tild_tile_range(uint32_t tile_x, uint32_t tile_y, uint64_t offset, uint32_t size);
-
-  size_t get_header_size() const;
-
-  uint64_t get_tile_offset(uint32_t idx) { return m_offsets[idx].offset; }
-
-  uint32_t get_tile_size(uint32_t idx) { return m_offsets[idx].size; }
-
-private:
-  /*
-   * Flags:
-   * bit 0   - dimensions 64 bit
-   * bit 1-2 - number of bits for offsets
-   * bit 3   - with tile sizes
-   * bit 4   - number of bits for size
-   * bit 5   - sequential ordering hint
-   * bit 6   - multidimensional (> 2D)
-   */
-  heif_tild_image_parameters m_parameters;
-
-  struct TileOffset {
-    uint64_t offset = TILD_OFFSET_NOT_LOADED;
-    uint32_t size = 0;
-  };
-
-  std::vector<TileOffset> m_offsets;
-
-  size_t m_header_size = 0;
-};
+class Image;
+class ImageOverlay;
 
 
 
@@ -228,314 +76,13 @@ public:
 
   Error read_from_memory(const void* data, size_t size, bool copy);
 
-  class Image : public ErrorBuffer
-  {
-  public:
-    Image(HeifContext* file, heif_item_id id);
-
-    ~Image();
-
-    void clear()
-    {
-      m_thumbnails.clear();
-      m_alpha_channel.reset();
-      m_depth_channel.reset();
-      m_aux_images.clear();
-    }
-
-    Error check_resolution(uint32_t w, uint32_t h) const {
-      return m_heif_context->check_resolution(w, h);
-    }
-
-    void set_resolution(int w, int h)
-    {
-      m_width = w;
-      m_height = h;
-    }
-
-    void set_primary(bool flag = true) { m_is_primary = flag; }
-
-    heif_item_id get_id() const { return m_id; }
-
-    //void set_id(heif_item_id id) { m_id=id; }  (already set in constructor)
-
-    // 32bit limitation from `ispe`
-    uint32_t get_width() const { return m_width; }
-
-    uint32_t get_height() const { return m_height; }
-
-    uint32_t get_ispe_width() const;
-
-    uint32_t get_ispe_height() const;
-
-    int get_luma_bits_per_pixel() const;
-
-    int get_chroma_bits_per_pixel() const;
-
-    Error get_preferred_decoding_colorspace(heif_colorspace* out_colorspace, heif_chroma* out_chroma) const;
-
-    bool is_primary() const { return m_is_primary; }
-
-    void set_size(int w, int h)
-    {
-      m_width = w;
-      m_height = h;
-    }
-
-    void process_before_write();
-
-    // -- thumbnails
-
-    void set_is_thumbnail()
-    {
-      m_is_thumbnail = true;
-    }
-
-    void add_thumbnail(const std::shared_ptr<Image>& img) { m_thumbnails.push_back(img); }
-
-    bool is_thumbnail() const { return m_is_thumbnail; }
-
-    const std::vector<std::shared_ptr<Image>>& get_thumbnails() const { return m_thumbnails; }
-
-
-    // --- alpha channel
-
-    void set_is_alpha_channel()
-    {
-      m_is_alpha_channel = true;
-    }
-
-    void set_alpha_channel(std::shared_ptr<Image> img) { m_alpha_channel = std::move(img); }
-
-    bool is_alpha_channel() const { return m_is_alpha_channel; }
-
-    const std::shared_ptr<Image>& get_alpha_channel() const { return m_alpha_channel; }
-
-    void set_is_premultiplied_alpha(bool flag) { m_premultiplied_alpha = flag; }
-
-    bool is_premultiplied_alpha() const { return m_premultiplied_alpha; }
-
-
-    // --- depth channel
-
-    void set_is_depth_channel()
-    {
-      m_is_depth_channel = true;
-    }
-
-    void set_depth_channel(std::shared_ptr<Image> img) { m_depth_channel = std::move(img); }
-
-    bool is_depth_channel() const { return m_is_depth_channel; }
-
-    const std::shared_ptr<Image>& get_depth_channel() const { return m_depth_channel; }
-
-
-    void set_depth_representation_info(struct heif_depth_representation_info& info)
-    {
-      m_has_depth_representation_info = true;
-      m_depth_representation_info = info;
-    }
-
-    bool has_depth_representation_info() const
-    {
-      return m_has_depth_representation_info;
-    }
-
-    const struct heif_depth_representation_info& get_depth_representation_info() const
-    {
-      return m_depth_representation_info;
-    }
-
-
-    // --- generic aux image
-
-    void set_is_aux_image(const std::string& aux_type)
-    {
-      m_is_aux_image = true;
-      m_aux_image_type = aux_type;
-    }
-
-    void add_aux_image(std::shared_ptr<Image> img) { m_aux_images.push_back(std::move(img)); }
-
-    bool is_aux_image() const { return m_is_aux_image; }
-
-    const std::string& get_aux_type() const { return m_aux_image_type; }
-
-    std::vector<std::shared_ptr<Image>> get_aux_images(int aux_image_filter = 0) const
-    {
-      if (aux_image_filter == 0) {
-        return m_aux_images;
-      }
-      else {
-        std::vector<std::shared_ptr<Image>> auxImgs;
-        for (const auto& aux : m_aux_images) {
-          if ((aux_image_filter & LIBHEIF_AUX_IMAGE_FILTER_OMIT_ALPHA) && aux->is_alpha_channel()) {
-            continue;
-          }
-
-          if ((aux_image_filter & LIBHEIF_AUX_IMAGE_FILTER_OMIT_DEPTH) &&
-              aux->is_depth_channel()) {
-            continue;
-          }
-
-          auxImgs.push_back(aux);
-        }
-
-        return auxImgs;
-      }
-    }
-
-
-    // --- metadata
-
-    void add_metadata(std::shared_ptr<ImageMetadata> metadata)
-    {
-      m_metadata.push_back(std::move(metadata));
-    }
-
-    const std::vector<std::shared_ptr<ImageMetadata>>& get_metadata() const { return m_metadata; }
-
-
-    // --- miaf
-
-    void mark_not_miaf_compatible() { m_miaf_compatible = false; }
-
-    bool is_miaf_compatible() const { return m_miaf_compatible; }
-
-
-    // === writing ===
-
-    void set_preencoded_hevc_image(const std::vector<uint8_t>& data);
-
-    const std::shared_ptr<const color_profile_nclx>& get_color_profile_nclx() const { return m_color_profile_nclx; }
-
-    const std::shared_ptr<const color_profile_raw>& get_color_profile_icc() const { return m_color_profile_icc; }
-
-    void set_color_profile(const std::shared_ptr<const color_profile>& profile)
-    {
-      auto icc = std::dynamic_pointer_cast<const color_profile_raw>(profile);
-      if (icc) {
-        m_color_profile_icc = std::move(icc);
-      }
-
-      auto nclx = std::dynamic_pointer_cast<const color_profile_nclx>(profile);
-      if (nclx) {
-        m_color_profile_nclx = std::move(nclx);
-      }
-    };
-
-    void set_intrinsic_matrix(const Box_cmin::RelativeIntrinsicMatrix& cmin) {
-      m_has_intrinsic_matrix = true;
-      m_intrinsic_matrix = cmin.to_absolute(get_ispe_width(), get_ispe_height());
-    }
-
-    bool has_intrinsic_matrix() const { return m_has_intrinsic_matrix; }
-
-    Box_cmin::AbsoluteIntrinsicMatrix& get_intrinsic_matrix() { return m_intrinsic_matrix; }
-
-    const Box_cmin::AbsoluteIntrinsicMatrix& get_intrinsic_matrix() const { return m_intrinsic_matrix; }
-
-
-    void set_extrinsic_matrix(const Box_cmex::ExtrinsicMatrix& cmex) {
-      m_has_extrinsic_matrix = true;
-      m_extrinsic_matrix = cmex;
-    }
-
-    bool has_extrinsic_matrix() const { return m_has_extrinsic_matrix; }
-
-    Box_cmex::ExtrinsicMatrix& get_extrinsic_matrix() { return m_extrinsic_matrix; }
-
-    const Box_cmex::ExtrinsicMatrix& get_extrinsic_matrix() const { return m_extrinsic_matrix; }
-
-
-    void add_region_item_id(heif_item_id id) { m_region_item_ids.push_back(id); }
-
-    const std::vector<heif_item_id>& get_region_item_ids() const { return m_region_item_ids; }
-
-    Error read_grid_spec();
-
-    bool is_grid() const { return m_is_grid; }
-
-    const ImageGrid& get_grid_spec() const { return m_grid_spec; }
-
-    const std::vector<heif_item_id>& get_grid_tiles() const { return m_grid_tile_ids; }
-
-
-    // --- tild
-
-    void set_tild_header(const TildHeader& header) { m_tild_header = header; m_is_tild = true; }
-
-    TildHeader& get_tild_header() { return m_tild_header; }
-
-    uint64_t get_next_tild_position() { return m_next_tild_position; }
-
-    void set_next_tild_position(uint64_t pos) { m_next_tild_position = pos; }
-
-  private:
-    HeifContext* m_heif_context;
-
-    heif_item_id m_id = 0;
-    uint32_t m_width = 0, m_height = 0;  // after all transformations have been applied
-    bool m_is_primary = false;
-
-    bool m_is_thumbnail = false;
-
-    std::vector<std::shared_ptr<Image>> m_thumbnails;
-
-    bool m_is_alpha_channel = false;
-    bool m_premultiplied_alpha = false;
-    std::shared_ptr<Image> m_alpha_channel;
-
-    bool m_is_depth_channel = false;
-    std::shared_ptr<Image> m_depth_channel;
-
-    bool m_has_depth_representation_info = false;
-    struct heif_depth_representation_info m_depth_representation_info;
-
-    bool m_is_aux_image = false;
-    std::string m_aux_image_type;
-    std::vector<std::shared_ptr<Image>> m_aux_images;
-
-    std::vector<std::shared_ptr<ImageMetadata>> m_metadata;
-
-    std::shared_ptr<const color_profile_nclx> m_color_profile_nclx;
-    std::shared_ptr<const color_profile_raw> m_color_profile_icc;
-
-    bool m_miaf_compatible = true;
-
-    std::vector<heif_item_id> m_region_item_ids;
-
-    bool m_has_intrinsic_matrix = false;
-    Box_cmin::AbsoluteIntrinsicMatrix m_intrinsic_matrix{};
-
-    bool m_has_extrinsic_matrix = false;
-    Box_cmex::ExtrinsicMatrix m_extrinsic_matrix{};
-
-    bool m_is_grid = false;
-    ImageGrid m_grid_spec;
-    std::vector<heif_item_id> m_grid_tile_ids;
-
-    bool m_is_tild = false;
-    TildHeader m_tild_header;
-    uint64_t m_next_tild_position = 0;
-  };
-
   Error check_resolution(uint32_t width, uint32_t height) const;
 
   std::shared_ptr<HeifFile> get_heif_file() const { return m_heif_file; }
 
   std::vector<std::shared_ptr<Image>> get_top_level_images() { return m_top_level_images; }
 
-  std::shared_ptr<Image> get_top_level_image(heif_item_id id)
-  {
-    for (auto& img : m_top_level_images) {
-      if (img->get_id() == id) {
-        return img;
-      }
-    }
-
-    return nullptr;
-  }
+  std::shared_ptr<Image> get_top_level_image(heif_item_id id);
 
   std::shared_ptr<const Image> get_top_level_image(heif_item_id id) const
   {
@@ -719,6 +266,8 @@ public:
 
   void write(StreamWriter& writer);
 
+  Error get_id_of_non_virtual_child_image(heif_item_id in, heif_item_id& out) const;
+
 private:
   std::map<heif_item_id, std::shared_ptr<Image>> m_all_images;
 
@@ -758,8 +307,6 @@ private:
                              std::shared_ptr<HeifPixelImage>& img,
                              const std::vector<uint8_t>& overlay_data,
                              const heif_decoding_options& options) const;
-
-  Error get_id_of_non_virtual_child_image(heif_item_id in, heif_item_id& out) const;
 };
 
 #endif
